source: share compact snapshots and remove scratch allocations
Share normalized supplier slices with Rc and narrow row indices to u32
within the validated token domain. Weighted entries shrink from 24 to 16
bytes on 64-bit targets without changing exact allocation arithmetic.

Freeze retained attribution buffers and power identities, but feed weight
projections directly into proportional allocation. Clone Strength
reductions only when an enemy Strength transition can change them.

Preserve normalization, frozen supplier ownership, capacity fallback,
ordering, and transactional updates. Independent allocation models and
seeded lifecycle scenarios cover the same invariants.

AI-Assisted: gpt-6-astra

+++ profiler-core/src/data/source/scenarios.rs
@@ -233,7 +233,7 @@ impl Scenario {
             .map(|share| {
                 let (name, slot) = match share.destination() {
                     Destination::Row(index) => {
-                        let row = &self.combat().cards[index];
+                        let row = &self.combat().cards[index as usize];
                         (row.id.as_str(), row.player)
                     }
                     Destination::Unknown(slot) => ("UNATTRIBUTED", slot),
